Add Movie List Component
A new component - MovieListComponent is created. This component
retrieves data from MovieListService and displays data in a Template.
The template use ngFor directive to process list of Movies.

A basic layout is added to the product. A css stylesheet is created to
control the look and feel of the layout.